chore: Move previews from Android packages to common packages
