+++ mockzilla-management-ui/mockzilla-desktop/build.gradle.kts
@@ -36,6 +36,7 @@ kotlin {
             implementation(compose.runtime)
             implementation(compose.material3)
             implementation(compose.components.resources)
+            implementation(compose.components.uiToolingPreview)
             implementation(compose.materialIconsExtended)
             implementation(libs.navigation.compose)
