+++ mockzilla-management-ui/mockzilla-management-ui-common/build.gradle.kts
@@ -49,6 +49,7 @@ kotlin {
             implementation(compose.runtime)
             implementation(compose.material3)
             implementation(compose.components.resources)
+            implementation(compose.components.uiToolingPreview)
             implementation(compose.materialIconsExtended)
             implementation(libs.navigation.compose)
 
@@ -125,6 +126,11 @@ kotlin {
     compilerOptions {
         freeCompilerArgs.addAll(CompilerConfig.freeCompilerArgs)
     }
+}
+
+dependencies {
+    /* Compose Previews */
+    debugImplementation(compose.uiTooling)
 }
 
 android {
